+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - AbstractControlBuilder added, related refactoring.
 - Control and ToggleControl.copyBuilder() added.
 - DefaultProgressWorkerDialogBuilder, bug fixed, did not set the locationRelativeTo for the progress dialog, it was always centered on application window. ProgressDialog bug fixed, didn't use the locationRelativeTo value provided by the builder.
+- Control and ToggleControl.copyBuilder() renamed copy().
 ### is.codion.framework.domain
 - IdentityKeyGenerator now throws exception in case the generated keys result set is empty.
 ### is.codion.framework.model